pyproject.toml
[tool.poetry]
name = "django-rest-framework-apis-in-a-nutshell"
version = "0.1.0"
description = "An API for a shoe store created with a Django REST framework and Django server."
authors = ["Kano Marvel <knmarvel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
flake8 = "^3.8.3"
djangorestframework = "^3.11.0"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
